projects/law_crm_project/requirements.txt
# Core Django
Django==4.2.7
psycopg2-binary==2.9.9
python-decouple==3.8

# Асинхронные задачи
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Аудит и безопасность
django-auditlog==2.3.0
django-guardian==2.4.0

# Работа с документами
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.0.0

# Поиск
django-filter==23.5

# API для интеграций
requests==2.31.0
beautifulsoup4==4.12.2

# Хранилище файлов
django-storages==1.14.2

# FSM для управления состояниями дел
django-fsm==2.8.1

# Утилиты
python-dateutil==2.8.2
phonenumbers==8.13.26
transliterate==1.10.2

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1

# WSGI сервер для production
gunicorn==21.2.0

# Работа с Excel (для экспорта отчетов)
openpyxl==3.1.2

# Валидация данных
django-phonenumber-field==7.2.0

